requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
_pytorch_select=0.2=gpu_0
absl-py=0.13.0=py37h06a4308_0
aiohttp=3.7.4=py37h27cfd23_1
appdirs=1.4.4=py_0
argcomplete=1.12.3=pypi_0
astroid=2.6.6=py37h06a4308_0
async-timeout=3.0.1=py37h06a4308_0
attrs=21.2.0=pyhd3eb1b0_0
audioread=2.1.9=py37h89c1867_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.6.0=pypi_0
beir=1.0.0=dev_0
blas=1.0=mkl
blinker=1.4=py37h06a4308_0
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.10.26=h06a4308_2
cached-property=1.5.2=py_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2021.10.8=py37h06a4308_0
cffi=1.14.6=py37h400218f_0
chardet=3.0.4=py37h06a4308_1003
click=8.0.1=pyhd3eb1b0_0
colorama=0.4.4=pypi_0
coverage=5.5=py37h27cfd23_2
cryptography=3.4.7=py37hd23ed53_0
cudatoolkit=11.1.74=h6bb024c_0
cycler=0.10.0=py37_0
cython=0.29.24=py37h295c915_0
dataclasses=0.8=pyh6d0b6a4_7
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=pypi_0
decorator=4.4.2=py_0
docker-pycreds=0.4.0=pypi_0
elasticsearch=7.9.1=pypi_0
entrypoints=0.3=pypi_0
eunjeon=0.4.0=pypi_0
expat=2.4.1=h2531618_2
faiss-cpu=1.7.2=pypi_0
ffmpeg=4.2.2=h20bf706_0
ffmpeg-python=0.2.0=py_0
filelock=3.0.12=pyhd3eb1b0_1
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=h5ab3b9f_0
future=0.18.2=py37_1
gettext=0.21.0=hf68c758_0
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
glib=2.69.1=h5202010_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=1.33.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.1=py_2
gputil=1.4.0=pyh9f0ad1d_0
grpcio=1.36.1=py37h2157cd5_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=3.2.1=py37h6c542dc_0
hdf5=1.10.6=hb1b8bf9_0
huggingface-hub=0.0.12=pypi_0
icu=58.2=he6710b0_3
idna=2.10=pyhd3eb1b0_0
imageio=2.9.0=py_0
imageio-ffmpeg=0.4.5=pyhd8ed1ab_0
importlib-metadata=3.10.0=py37h06a4308_0
importlib_metadata=3.10.0=hd3eb1b0_0
iniconfig=1.1.1=pyhd3eb1b0_0
install=1.3.5=pypi_0
ipykernel=6.6.0=pypi_0
ipython=7.30.1=pypi_0
isort=5.9.2=pyhd3eb1b0_0
jedi=0.18.1=pypi_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
jpype1=1.3.0=pypi_0
jsonlines=3.0.0=pypi_0
jupyter-client=7.1.0=pypi_0
jupyter-core=4.9.1=pypi_0
kiwisolver=1.3.1=py37h2531618_0
konlpy=0.5.2=pypi_0
lame=3.100=h7b6447c_0
lazy-object-proxy=1.6.0=py37h27cfd23_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libblas=3.9.0=1_h6e990d7_netlib
libcblas=3.9.0=3_h893e4fe_netlib
libffi=3.3=he6710b0_2
libflac=1.3.3=h9c3ff4c_1
libgcc-ng=11.1.0=hc902ee8_8
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libidn2=2.3.2=h7f8727e_0
liblapack=3.9.0=3_h893e4fe_netlib
libllvm10=10.0.1=hbcb73fb_5
libogg=1.3.5=h27cfd23_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.17.2=h4ff587b_1
librosa=0.8.1=pyhd8ed1ab_0
libsndfile=1.0.31=h9c3ff4c_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libvorbis=1.3.7=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lime=0.2.0.1=pypi_0
llvm-openmp=12.0.1=h4bd325d_1
llvmlite=0.36.0=py37h612dafd_4
lxml=4.7.1=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py37h06a4308_0
matplotlib=3.4.2=py37h06a4308_0
matplotlib-base=3.4.2=py37hab158f2_0
matplotlib-inline=0.1.3=pypi_0
mccabe=0.6.1=py37_1
mecab-python=0.996-ko-0.9.2=pypi_0
mkl=2019.5=281
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.3.0=py37h54f3939_0
mkl_random=1.0.4=py37hd81dba3_0
more-itertools=8.8.0=pyhd3eb1b0_0
moviepy=1.0.1=py_0
multidict=5.1.0=py37h27cfd23_2
munkres=1.1.4=py_0
mypy=0.910=pyhd3eb1b0_0
mypy_extensions=0.4.3=py37_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.4=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=2.6.3=pypi_0
ninja=1.10.2=hff7bd54_1
nltk=3.6.2=pyhd3eb1b0_0
numba=0.53.1=py37ha9443f7_0
numpy=1.21.4=pypi_0
oauthlib=3.1.1=pyhd3eb1b0_0
olefile=0.46=py37_0
openh264=2.1.0=hd408876_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1l=h7f8727e_0
packaging=21.0=pyhd3eb1b0_0
pandas=1.2.4=py37ha9443f7_0
parso=0.8.3=pypi_0
pathspec=0.7.0=py_0
pathtools=0.1.2=pypi_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.3.1=py37h2c7a002_0
pip=21.2.2=py37h06a4308_0
pluggy=0.13.1=py37h06a4308_0
pooch=1.4.0=pyhd3eb1b0_0
pprintpp=0.4.0=py37h89c1867_2
proglog=0.1.9=py_0
promise=2.3=pypi_0
prompt-toolkit=3.0.23=pypi_0
protobuf=3.17.2=py37h295c915_0
psutil=5.8.0=py37h27cfd23_1
ptyprocess=0.7.0=pypi_0
py=1.10.0=pyhd3eb1b0_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycodestyle=2.7.0=pyhd3eb1b0_0
pycparser=2.20=py_2
pydantic=1.8.2=py37h27cfd23_0
pygments=2.10.0=pypi_0
pyjwt=2.1.0=py37h06a4308_0
pylint=2.9.6=py37h06a4308_1
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h05f1152_2
pysocks=1.7.1=py37_1
pysoundfile=0.10.3.post1=pyhd3deb0d_0
pytest=6.2.4=py37h06a4308_2
pytest-cov=2.12.1=pyhd3eb1b0_0
python=3.7.11=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.7=2_cp37m
pytorch=1.8.2=py3.7_cuda11.1_cudnn8.0.5_0
pytrec-eval=0.5=pypi_0
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.2.0=pypi_0
pyyaml=5.4.1=py37h27cfd23_1
pyzmq=22.3.0=pypi_0
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
regex=2021.4.4=py37h27cfd23_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
resampy=0.2.2=py_0
rsa=4.7.2=pyhd3eb1b0_1
sacremoses=0.0.43=pyh9f0ad1d_0
scikit-image=0.19.0=pypi_0
scikit-learn=0.24.2=py37ha9443f7_0
scipy=1.6.2=py37h91f5cce_0
sentence-transformers=2.2.0=pypi_0
sentencepiece=0.1.96=pypi_0
sentry-sdk=1.5.12=pypi_0
setproctitle=1.2.2=py37h27cfd23_1004
setuptools=52.0.0=py37h06a4308_0
shortuuid=1.0.9=pypi_0
sip=4.19.8=py37hf484d3e_0
six=1.16.0=pyhd3eb1b0_0
smmap=5.0.0=pypi_0
sqlite=3.36.0=hc218d9a_0
tbb=2020.3=hfd86e86_0
tensorboard=2.5.0=py_0
tensorboard-plugin-wit=1.6.0=py_0
threadpoolctl=2.2.0=pyhbf3da8f_0
tifffile=2021.11.2=pypi_0
tk=8.6.10=hbc83047_0
tokenizers=0.10.1=py37hcb7a40c_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=0.8.2=py37
torchtext=0.6.0=py_1
torchvision=0.9.2=py37_cu111
tornado=6.1=py37h27cfd23_0
tqdm=4.62.0=pyhd3eb1b0_1
traitlets=5.1.1=pypi_0
transformers=4.9.1=pyhd8ed1ab_0
tweepy=4.4.0=pypi_0
typed-ast=1.4.3=py37h7f8727e_1
typing-extensions=3.10.0.0=hd3eb1b0_0
typing_extensions=3.10.0.0=pyh06a4308_0
urllib3=1.26.6=pyhd3eb1b0_1
wandb=0.12.16=pypi_0
wcwidth=0.2.5=pypi_0
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.12.1=py37h7b6447c_1
x264=1!157.20191217=h7b6447c_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yarl=1.6.3=py37h27cfd23_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
